requirements.txt
futu-api
psycopg2-binary
pandas
